+++ src/api/get-coin-info-by-id.js
@@ -1,19 +1,9 @@
+import GetCoinInfo from './get-coin-info.js';
+
 /**
- * @param {MinterApiInstance} apiInstance
- * @return {function(*): (Promise<CoinInfo>)}
+ * @deprecated
+ * @type {function(MinterApiInstance): function((string|number), AxiosRequestConfig=): Promise<CoinInfo>}
  */
-export default function GetCoinInfoById(apiInstance) {
-    /**
-     * Get nonce for new transaction: last transaction number + 1
-     * @param {number|string} coinId
-     * @param {AxiosRequestConfig} [axiosOptions]
-     * @return {Promise<CoinInfo>}
-     */
-    return function getCoinInfoById(coinId, axiosOptions) {
-        return apiInstance.get(`coin_info_by_id/${coinId}`, axiosOptions)
-            .then((response) => {
-                response.data.id = Number(response.data.id);
-                return response.data;
-            });
-    };
-}
+const GetCoinInfoById = GetCoinInfo;
+
+export default GetCoinInfoById;