+++ test/api/get-coin-info.test.js
@@ -32,3 +32,32 @@ describe('GetCoinInfo', () => {
             });
     }, 30000);
 });
+
+describe('GetCoinInfo by ID', () => {
+    test('should work gate', () => {
+        expect.assertions(1);
+
+        return minterGate.getCoinInfo(0)
+            .then((coinInfo) => {
+                expect(coinInfo.symbol === 'MNT' || coinInfo.symbol === 'BIP').toBeTruthy();
+            })
+            .catch((error) => {
+                logError(error);
+                throw error;
+            });
+    }, 30000);
+
+    test('should work node', () => {
+        expect.assertions(1);
+
+        return minterNode.getCoinInfo(0)
+            .then((coinInfo) => {
+                expect(coinInfo.symbol === 'MNT' || coinInfo.symbol === 'BIP').toBeTruthy();
+            })
+            .catch((error) => {
+                logError(error);
+                throw error;
+            });
+    }, 30000);
+});
+